move GoogleMapsEmbed component out of locations folder
